Do not unprotect quoted strings and quote filename in shell commands.
First the filename in shell commands must be quoted to avoid any
translation. Second we do not want to unprotect shell arguments when
they are quoted. Both of these changes are needed to properly support
Windows full path names which contain back-slashes that we do not want
to interpret as escape sequences.
--
This fixes support of the "Add Annotations" GPS action on Windows.
--
Fixes H626-028.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@126108 936e1b1b-40f2-da11-902a-00137254ae57