Do not use __gnat_stat() which is using support for large files.
We do not want to immort this routine there as some platforms are
now using the version supporting large files. The advantage of using
was for the Win32 support but GNATCOLL is not supporting Set_Readable
on Win32 anyway.

If large file support is needed then it will have to be implemented
into GNATCOLL directly at this point.
--
Fixes I323-003 regression.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@141837 936e1b1b-40f2-da11-902a-00137254ae57