(Library_Files): fix wrong computation of project controlling an ALI file
As a result, all the ALI files for xmlada were thought to be in the
project xmlada_sax.gpr, whereas schema.ads for instance should be part
of xmlada_schema.gpr. This occurred when an object directory was shared
between projects, and impacting file name resolution when parsing
ALI files.
Regression introduced in r223417 on Feb 17th when adding aggregate
projects support in gnatinspect.

N304-018

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@224030 936e1b1b-40f2-da11-902a-00137254ae57