protect against corrupted sqlite db on startup
When the startup is corrupted enough that 'pragma user_version' fails,
we simply want to recreate it from scratch. This used to raise a
Constraint_Error.
There are other cases of corruption which do not impact the startup
but will make later queries fail. These will be handled in later
check ins.
Part of N224-020

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@224047 936e1b1b-40f2-da11-902a-00137254ae57