Avoid exceptions when computing contexual menu with aggregate projects
This check in fixes a number (but not all) of calls to GNATCOLL.Projects.Info
which are invalid in the context of aggregate projects. The project
information is either read from the source view, or in some cases we can
simply take the first available project since that does not impact the
module (VCS for instance).

With this change, the contextual menu now properly displays with
aggregate projects.

source views are now associated with a project, although for now it is
simply the first possible project. This is the first part for
N305-035

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@224088 936e1b1b-40f2-da11-902a-00137254ae57