(Imports, Imported_By): fix queries
These queries were returning correct file names, but always associated
with the project of the file we started from. This was apparently
broken when we added support for aggregate projects, so was never
released in an official release.
This broke the computation in the File Dependency browser.

Part of N513-005