Do not try to spawn <target>-gnatls when target comes from the .cgpr file
Otherwise, we can end up with the following scenario: gnatcoll spawns
gprconfig (when Set_Automatic_Config_File is spawned) with no specific
target mentioned. The resulting auto.cgpr contains an attribute Target
set to the native platform (e.g. "x86-windows"). This is then used by
the project manager to initialize the Target attribute in the view of
the project and therefore we would try to spawn x86-windows-gnatls.

Detected as part of NA15-045, although has no effect on the list of
sources

Change-Id: Idca3d381b7505ffdee9b51cbf7b9a3fb27d219e4