Do not add projects with no view to project iterators
If for some reason the view has not been computed for given project
it won't be added to the iterator during Start and Start_Reversed.

Even before the view is computed for project tree, a root aggregate
project has partial view, consisting ofthe list of aggregated projects.
That caused Start and Start_Reversed to begin the iteration
but then crash due to abcense of the view of aggregated projects.

This also affected Object_Path. Object_Path now returnes epmty path
if the view has not been recomputed.

Comments added to Object_Path and Start.

for NA08-021

Change-Id: Ifa7a1bb373779686787055e40a10a5114724df2c