Fix filename normalization issues when name starts with "../.."
In this case, the first ".." is inserted in the output string (since there
is no known parent dir), and the second one removes it (assuming the
preceding dir is a valid directory).
Fixes regression in O618-010.
This bug has been around for a long time, but was made more visible by the
recent fix 6d7d020dbea41365486707c35ff7fbcddc85468a

Change-Id: I5245afc722998afa63935db10456168e04335eba