(Find_Stream): Fix expansion of $$
Environment variable expansion was recently improved in stream names,
however this introduced a regression due to the fact that a variable
whose name is the delimiter (i.e. $$ here) is handled by GNATCOLL.Templates
only if provided as a static substring (not through the callback).

Fixes O706-030 regression introduced by change for O523-004

Change-Id: Id490025bf684e16f178d8b6a8b9bf90d2afb0128