(gnatinspect): support for --lang
This commit adds support for a new command line switch (--lang) that
can be used to specify custom naming schemes for languages. In particular,
users can now define, in GPS, new languages based on Ada (that also
generate a .ali file or similar) and have cross-reference information
for them. The preferred approach, though, is to use a configuration
project file.

M304-034 (3/3)

Change-Id: I9d8d312e51756d154dce88980ff031f2d31b7412