Iconv: export a symbol which was missing from --without-iconv builds
For --without-iconv builds, GNATCOLL.Iconv is nevertheless available and
is supposed to be a mere stub. However this stub was missing a C symbol;
as a result, programs using the corresponding procedure would not build
properly. This change adds a function that exports this symbol in order
to fix libiconv-less builds.

Change-Id: I062f1c2f6829d01860d0a5f31ba6287b0bf2ea7e
TN: OA27-033