Make Index matter in Has_Attribute
Fix the fact that Has_Attribute returns true for given
Index if there is the same attribute with just any index,
matching the given one or not.
Also update the comments on the head of Attribute section
to mention that language names might be case-sensitive
on some platforms.

for OC17-040

Change-Id: I8977e6a9fa28bd704b51cd9844294e97aec4eb1e