Add support for substrings
Substrings do not require any memory allocation or copy of the
characters, when using copy on write.
We need to store one additional integer for big strings, which
means that the optimal size for small strings is now 23 instead of
15 (which provides its own speedup since a larger fraction of
strings fit as small strings).

We achieve the following speedup, in a test that repeatedly take
substrings:

========== Unbounded strings =============
Trim                   : 0.368205000s 100%
======== ssize= 23 COW=FALSE ========
Trim                   : 0.179822000s  49%
======== ssize= 23 COW=TRUE ========
Trim                   : 0.067753000s  18%
======== ssize= 127 COW=TRUE ========
Trim                   : 0.053445000s  15%

Q216-036

Change-Id: I1b719adf6ad4ce430760e2c5a7ed349bdf165689