Q304-013: Add a temporary workaround
Don't call GNATCOLL.Projects.Executable_Name with Include_Suffix set
to True for non-native targets, unless the Builder'Executable_Suffix
has been explicily set in the root project.

This is needed to avoid retrieving an excutable name with the .exe
suffix on Windows for non-native projects.

This workaround should be removed once the refactoring of
GNATCOLL.Projects is done. 

Change-Id: I8337398efff6f07c6f5f6c3bbdafbff6ff82dbe7