QC22-014 extra error message for undefined externals
In IDE mode if there was an error message about an undefined
external reference issue an extra error message explaining
that project has been loaded incompletely and needs to be
reloaded after corresponding externals are set.

This change has already been merged to gnatcoll,
adding it to the new repo to provide consistent behaviour.

Change-Id: I9ce7eb60908fa0f8f2847e18c26684e771c0a897