R115-026 fix Remove_Imported_Project
When removing an imported project and it no longer
is used anywhere in the project tree, remove it also
from the htable in GPR.Tree so that a different
project with same name can be imported afterwards.

This change has already been merged to gnatcoll,
adding it to the new repo to provide consistent behaviour.

Change-Id: I1a3aee5ca0acd3358cd53b2461e95e01d3e19667