GNATCOLL.JSON: add Read overloads to return error information
The current Read functions return a JSON_Value on success while they
raise an exception and display information about the parsing error on
the standard output.

This patch introduces two new overloads to Read that return instead a
discriminated record: Read_Result, which contain the JSON_Value when
parsing was successful and which contain instead the information about
the error on failure. This API is more friendly with tools that want to
keep control on their output stream: graphical programs, servers, ...

Change-Id: I96b1f3614ede44fb026552d8aa16fe7155b15018
TN: RA19-011