S528-022 Write temporary files in the temporary directory
When writing a file using the GNATCOLL.VFS.Write API, if the
file already exists and we're not in Append mode, we write
to a temporary file, for safety.

With this patch, the temporary file gets created in the
temporary directory. If this name is conflicting, try
with 10 more names before giving up.

This is to increase the reliability for using Write on
slow or networked filesystems, and to avoid leaving
stray temporary files in user directories in case the program
using this API crashes before deleting the temporary
file.


Slightly improve error reporting in the Close procedure.

Change-Id: I69f4315a48f7cbd1f67e2dd41219c0a043c80cdd