Set target and runtime from config file
Add Set_Target_And_Runtime_From_Config query to API that sets
values of Target and Runtime and also selects a proper gnatls
based on the path to ada driver from the configuration file.
Call this query during load if the config file has been specified
and exists.

for T507-023

Change-Id: Ieac4fd939be65cf278ed9148e24fa42e91d66457