T807-007 add target normalization
When possible and gprconfig KB is on the path, parse targetset.xml
at the beginning of first project load to populate normalization
dictionary. When comparing given target against the list of possible
host names normalize the target first. This allows to not add target
specific prefix to gnatls on native platforms when forced target comes
in a non-canonical form.

Change-Id: Ide6cbc1db6d5cb78a2a001f47b60a47e2325286e