GNATCOLL.Memory: Improve documentation of the Stack_Trace_Depth parameter
Setting this parameter to zero when calling the Configure procedure
has the effect of disabling completely the tracking of the biggest
memory users, which wasn't clear from the current documentation.
So this patch enhances the documentation of both the Configure
procedure as well as the Dump procedure to make that explicit.

Change-Id: Ib8c4733b9e821779cf52e5bfc892969fddea44b5
TN: V302-002