Store test output in file with testsuite basic driver
Store test output in a file named "test_results.out" in the test
working directory. It can be useful for post-test Python scripts
that need to process the output.

Note that a similar approach could be done with the build_run_diff
testsuite driver. However, contrary to the basic driver case,
Python files are copied into the test working directory. It can be
problematic if the pre- and post-scripts need parent Python scripts.
Even soft symbolic links shall not be used in this case, as the
script copy is made with e3.fs.sync_tree, which copies links but
not their target. So using relative target paths for parent directories
would not work.

TN: W329-010
Change-Id: Ie8f4ba08cba6892c7e88cefc86925ca44e2bdb12