Implement Symbolic_Link_Is_Internal
The symlink / symbolic link is a special file, whose purpose is to point
to a file or a directory, called the target.

Symbolic_Link_Is_Internal indicates if a symbolic link target is external
to the specified base directory. A link target is external in
several cases:
- Link target is an absolute path.
- Link target is a relative path pointing directly out of the
base directory.
- Link target is a relative path which leaves the base
directory at least once.

Not supported on Windows.

Tests have been implemented.

Issue: #14

TN: W329-010
Change-Id: I3c5cf798d12f5196e12af37241cc3029ec42f054