testsuite: add copy_files_on_target option to run_cross call
Some tests have data files which need to be copied on target.
The files are listed in a "data" field in test.yaml. Check for
the data field and read its contents to copy_files_on_target.
This allows run_cross' ssh.py to handle copying the given files.

TN: T616-039