Update instructions explaining the use of GNATcoll.Memory
The part explaining how to incorporate an alternative body
for s-memory was obsolete, referring to the old gnatmake -a
mechanism. Describe how to achieve that effect with an
additional project file instead, from which we can control
the compilation switches to add the required -gnatg.

Fix the proposed alternative s-memory.adb to prevent an
error on the "end;" line, where -gnatg requires the package
name.