Fix coverage build
If the gnatcov_rts.gpr installed by gnatcov setup is not found by
gnatcov instrument on the GPR_PROJECT_PATH, gnatcov instrument assumes
that the dump-trigger mode is manual, which triggers a bug. This bug
should be fixed, but the dump-trigger mode for gnatcoll-core should not
be manual, so fix this to avoid triggering the bug.