Fix test coverage visualization for gitlab MR's "changes" tab
The coverage generation of the gnatcoll-core's CI rely on `anod`, more specifically
`anod build gnatcoll-core -Qcoverage` and `anod test gnatcoll-core -Qcoverage`.
The latter will run `run-tests` that generates the coverage files thanks to the
`gnatcov coverage` command. The cobertura.xml file is used by the CI to show
coverage on MR's "changes" tab. It did not work before this modification,
because the sources paths were not relative to the gnatcoll-core repository,
but absolute to the gitlab's runner directory.

To remove this absolute prefix, --source-root can be given to the `gnatcov
coverage` command. This commit adapts `run-tests`, so it can accept this new
option that will be passed to `gnatcov`.

Co-Authored-By: Jerome Lambourg <lambourg@adacore.com>
Issue: #26
Depends-On: eng/shared/anod!2607