Add "assert and diff" testsuite driver
basic.py driver is designed to check for assertion, while build_run_diff.py
focuses on the test output. In the context of Sync_Trees testing, both
functionalities are relevant.

This commit merges the two drivers into one and also improves the "data"
functionality, so data can be copied from the test working directory:
in the context of cross testing, data can be generated dynamically with
a pre_test.py script in the host working directory and copied to the target's
one.

Issue: eng/toolchain/gnatcoll-core#30

Change-Id: If90d25e737339a53837aec69f6fd7d45a8443c16