Fix sync_trees cross tests
The cross testing did not work for several reasons:
- the source and destination directories generated by pre_test.py were
  not copied to the target working directory
- the check of sync_trees results done in post_test.py was not possible,
  as post_test.py is executed in the host working directory, while
  sync_trees results were available only in the target's one.

The first issue was partially fixed in a previous commit with the addition of
the "assert and diff" testsuite driver : the "data" field in test.yaml can copy
data from the test working directory and the test source directory. This allows
the source and destination directories generated by pre_test.py (so generated
in the test working directory) to be copied to the target working directory.

The second issue was a too complex to solve, so we do not rely anymore on
post_test.py to check trees content. Instead, src and dst trees are displayed
by the Ada test executable, and checked afterwards by the testsuite driver.

Symbolic links are complex to manage in a cross environment, so tests
including them are skipped.

Issue: eng/toolchain/gnatcoll-core#30

Change-Id: If0e08c19c51f4fe60f4f0d1fc55e6b78b8efbb5f