Do not replace the T delimiter with a space to respect the ISO 8601 norm
From Wikipedia:
"""
A single point in time can be represented by concatenating a complete date
expression, the letter "T" as a delimiter, and a valid time expression.
For example, "2007-04-05T14:30".
Separating date and time parts with other characters such as space is not
allowed in ISO 8601, but allowed in its profile RFC 3339.
"""

The "T" delimiter is now correctly supported by GNAT.Calendar.Time_IO.