Preserve trailing directory separator with Rename_And_Move
This commit is the continuity of the work from Charles Pigott (LordAro) that
can be found here: https://github.com/AdaCore/gnatcoll-core/issues/32.

When calling GNATCOLL.Projects.Rename_And_Move, if you pass in a directory
that is missing the trailing slash (such as a path generated by
Ada.Directories.Containing_Directory), the resulting renamed gpr file is
missing path separators in attributes such as Object_Dir, which can lead to
invalid directory or pathnames.