Fix assert traces in case of success
"Location" was prefixed to the success message with GNATCOLL.Traces.Assert
if the colors were enabled.
This was caused by the wrong arguments given to the Trace procedure: "Location"
was given to the Color argument, which prefixes the success message when colors
are enabled.

This commit also simplifies the debug_config_parsing tests.