Return '__others__' as the name for the 'others' GPR special attribute index
The term "others" (without quotes) refers to a special attribute index used to
specify all the indexes existing in the project. The Attribute_Indexes operates
with strings, while the GPR API works with Name_Ids in this case.

A special value, "__others__", is now returned so that GNATCOLL users can
differentiate the "others" index from the previously mentioned special one.