Opt_Parse: refactor to avoid a new accessibility check failure
Use cursor types rather than reference types for return types when the
referenced vector element come from a function formal. This is necessary
to avoid the failure of a new accessibility check that was just
implemented in GNAT. It will also be necessary once reference types are
made limited (as per AI22-0082).