Prevent infinite loop in wait_for_processes__unix
If the timeout passed by Wait_For_Processes to Wait_For_Sigchld was
infinite (i.e., -1), then:

effective_timeout was neither initialized nor used in the loop condition.
As a result, the loop was never entered, and the function always returned -1.

This commit also fixes the uninitialized tv value, which could lead to
undefined behavior when using infinite timeouts.

Fixes: eng/toolchain/gnatcoll-core#138