Refactor testsuite code
Rewrite the root run-tests python script to share the same properties as
with sub run-tests scripts. The common properties are stored in a new
file common_testsuite.py, that inherits from gprproject testsuite.

Also, remove th usage of the drivers that are already defined in
gprproject, like basic.py, and the functions coming dedicated to gnatcov and
valgrind.