Testsuite: Fix missing argument to gnatcov instrument call
A recent change adapted the coverage build of GNATCOLL to adapt to
a change in how gnatcov accepts the `--relocate-build-tree` switch.
It now systematically requires an argument. The previous change failed
to update the call to gnatcov in the testsuite drivers, thus resulting
in empty coverage reports.

This change remediates this issue.